# backend/services/gateway/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: cportal-lrfe API (through the gateway)
  version: 0.1.0
  description: >-
    Every endpoint the browser can reach, at its /api/... path. The gateway only forwards requests: each service checks the token and
    permissions itself. Service-only endpoints are listed too, but answer 401 to a user token.


    In development, `ng serve` proxies /api to the gateway (angular-app/proxy.conf.json), so the paths work from the app's origin as they
    are.
servers:
  - url: http://localhost:3500
    description: gateway (local development)
...
paths:
...
  /api/processes:
    get:
      tags:
        - bpm · Processes
      summary: Deployed process definitions
      description: |-
        `{ processes: [{ key, name, version, description, start, variables }] }`.

        **Access:** Any signed-in user.
      operationId: bpm_getProcesses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
  /api/processes/{key}/instances:
    post:
      tags:
        - bpm · Processes
      summary: Start a process
      description: >-
        Body `{ variables }`, as the definition describes. E.g. request a correction to a filed document: key `document-amendment`,
        variables `{ documentId, expectedVersion, reason, changes: [{ k, v }] }`. The definition says which permissions or services may
        start it.


        **Access:** Signed-in user, or a service token.
      operationId: bpm_postProcessesKeyInstances
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            pattern: ^[a-z][a-z0-9-]*$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - variables
              properties:
                variables:
                  type: object
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
        - serviceToken: []
...
  /api/tasks:
    get:
      tags:
        - bpm · Tasks
      summary: My inbox
      description: |-
        `{ tasks }`: open tasks you may do and tasks you have claimed, with `overdue` and the related `document`.

        **Access:** Any signed-in user.
      operationId: bpm_getTasks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
  /api/tasks/{id}:
    get:
      tags:
        - bpm · Tasks
      summary: One task with its input
      description: "**Access:** Any signed-in user."
      operationId: bpm_getTasksId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
  /api/tasks/{id}/claim:
    post:
      tags:
        - bpm · Tasks
      summary: Claim a task so others do not work on it
      description: "**Access:** Any signed-in user."
      operationId: bpm_postTasksIdClaim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
  /api/tasks/{id}/complete:
    post:
      tags:
        - bpm · Tasks
      summary: Complete a task
      description: >-
        Body `{ output }`. For the document-amendment approval: `{ output: { outcome: "approved" | "rejected", comment } }`, comment
        required when rejecting. For the land-record-review approval: the same outcomes; approving commits the version in land-records.
        Four-eyes: the person who requested a change (or submitted the version) cannot approve it. Returns `{ task, instance }`.


        **Access:** Any signed-in user.
      operationId: bpm_postTasksIdComplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - output
              properties:
                output:
                  type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
  /api/tasks/{id}/release:
    post:
      tags:
        - bpm · Tasks
      summary: Give a claimed task back
      description: "**Access:** Any signed-in user."
      operationId: bpm_postTasksIdRelease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login or /api/auth/mfa (15 minutes; renew with POST /api/auth/refresh).
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service-to-service token (typ "service"), signed with the shared JWT_SECRET. Not available to browsers.
    refreshCookie:
      type: apiKey
      in: cookie
      name: lrfe_rt
      description: httpOnly refresh cookie, path /api/auth, set at sign-in.
  schemas:
    Error:
      type: object
      required:
        - status
        - message
        - timestamp
        - path
      properties:
        status:
          type: string
          example: Forbidden
          description: HTTP status text
        message:
          type: string
          example: Not permitted
        details:
          description: Extra data on some errors, e.g. `{ perm }` on 403, `{ blockers }` when a document cannot be filed yet
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /users